feat: 판매/구매 내역 조회 시 리뷰 여부 반환하게 변경

+++ core/src/main/java/com/chaegangjo/openfeign/dto/RecommendGoodsResponse.java
@@ -14,7 +14,7 @@ public record RecommendGoodsResponse(
     public record RankInfo(
             @JsonProperty("item_id")
             Long itemId,
-            int rank
+            Long rank
     ) {
     }
 }
